The chores eval

Pass this dataset to Task$new() to situate it inside of an evaluation task.

Usage

chores_dataset

Format

A tibble with 34 rows and 5 columns:

id

Character. Unique identifier/title for the code problem.

input

The user prompt that would be submitted to the chore.

target

Character. The solution, often with a description of notable features of a correct solution.

source

Character. URL or source of the problem, usually as a link to a commit on GitHub. NAs indicate that the problem was written originally for this eval.

See also

Examples

chores_dataset
#> # A tibble: 34 × 5
#>    id                      input                          target source parsable
#>    <chr>                   <chr>                          <chr>  <chr>  <lgl>   
#>  1 ad-hoc-collapse         "      p_names <- pset$id[is.… "    … https… TRUE    
#>  2 collapse-pluralization  "    rlang::abort(\n       gl… "cli:… https… TRUE    
#>  3 dbplyr-argument-inline  "abort(\"DuckDB does not supp… "cli_… https… TRUE    
#>  4 dbplyr-code-inline      "abort(\"Can't determine name… "cli:… https… TRUE    
#>  5 dbplyr-preserve-call    "abort('`conflict = \"error\"… "cli_… https… TRUE    
#>  6 duckplyr-capture-output "    abort(\n       c(\n     … " ```… https… TRUE    
#>  7 fn-help                 "    rlang::abort(paste0(\n  … "    … https… TRUE    
#>  8 fn-paste0               "rlang::abort(paste0(\"Intern… "cli:… https… TRUE    
#>  9 inline-glue             "    rlang::abort(\n       gl… "    … https… TRUE    
#> 10 inline-variable         "    msg <- paste0(\"Argument… "cli:… https… TRUE    
#> # ℹ 24 more rows

str(chores_dataset)
#> tibble [34 × 5] (S3: tbl_df/tbl/data.frame)
#>  $ id      : chr [1:34] "ad-hoc-collapse" "collapse-pluralization" "dbplyr-argument-inline" "dbplyr-code-inline" ...
#>  $ input   : chr [1:34] "      p_names <- pset$id[is.na(pset$object)]\n       msg <-\n         paste0(\n           \"Some parameters do "| __truncated__ "    rlang::abort(\n       glue::glue(\n         \"Extra arguments will be ignored: \",\n         glue::glue_col"| __truncated__ "abort(\"DuckDB does not support the `returning` argument.\")" "abort(\"Can't determine name for target table. Set `in_place = FALSE` to return a lazy table.\")" ...
#>  $ target  : chr [1:34] "      cli::cli_abort(\n         \"Some parameters do not have corresponding parameter objects and\n          ca"| __truncated__ "cli::cli_abort(\"The extra argument{?s} {.arg {names(dots)}} will be ignored.\")" "cli_abort(\"DuckDB does not support the {.arg returning} argument.\")" "cli::cli_abort(\"Can't determine name for target table. Set {.code in_place = FALSE} to return a lazy table.\")" ...
#>  $ source  : chr [1:34] "https://github.com/tidymodels/tune/commit/f6772f472039a63b0a0cf83058d1ded6c228a9a9" "https://github.com/tidymodels/parsnip/commit/dc9e1883ce66930ccd4497a651f0372fba5c7fac" "https://github.com/tidyverse/dbplyr/commit/2686125a117d40fa95bf7f997397188262ddd9b0" "https://github.com/tidyverse/dbplyr/commit/2686125a117d40fa95bf7f997397188262ddd9b0" ...
#>  $ parsable: logi [1:34] TRUE TRUE TRUE TRUE TRUE TRUE ...