The chores solver

Pass this function to Task$new() as the solver to process input prompts from the chores dataset with a specified language model.

Usage

chores_solver(inputs, ..., solver_chat)

Arguments

inputs

Character vector of user prompts, likely from chores_dataset$input.

...

Additional arguments passed to the chat_parallel method.

solver_chat

An ellmer chat object to use for solving the prompts. This must be a Chat object from the ellmer package.

Value

A list with the following components:

result

Character vector of model responses, one for each input.

solver_chat

List of Chat objects used to generate each response, with the same length as inputs.

See also